src/topprism_chatopt/requirements.txt
streamlit==1.37.0
pandas==2.2.0
ortools==9.10.4067
sentence-transformers==2.5.1
faiss-cpu==1.8.0
openai==1.35.0
plotly==5.18.0
numpy==1.26.0